Avoid more conversion error in geometry section

+++ codes/Geometry/FarthestPair.cpp
@@ -1,5 +1,5 @@
 // p is CCW convex hull w/o colinear points
-int n = p.size(), pos = 1; lld ans = 0;
+int n = (int)p.size(), pos = 1; lld ans = 0;
 for (int i = 0; i < n; i++) {
   P e = p[(i + 1) % n] - p[i];
   while (cross(e, p[(pos + 1) % n] - p[i]) >
